Move charset conversion for MIDI import to separate files
