+++ mscore/importmidi/importmidi_lyrics.cpp
@@ -1,6 +1,7 @@
 #include "importmidi_lyrics.h"
 #include "importmidi_inner.h"
 #include "importmidi_fraction.h"
+#include "importmidi_charset.h"
 #include "importmidi_chord.h"
 #include "importmidi_operations.h"
 #include "libmscore/box.h"
