+++ mtest/CMakeLists.txt
@@ -77,6 +77,7 @@ add_library(
       ${PROJECT_SOURCE_DIR}/mscore/importmidi/importmidi_simplify.cpp
       ${PROJECT_SOURCE_DIR}/mscore/importmidi/importmidi_voice.cpp
       ${PROJECT_SOURCE_DIR}/mscore/importmidi/importmidi_key.cpp
+      ${PROJECT_SOURCE_DIR}/mscore/importmidi/importmidi_charset.cpp
       ${PROJECT_SOURCE_DIR}/mscore/exportmidi.cpp
       ${PROJECT_SOURCE_DIR}/mscore/importxml.cpp
       ${PROJECT_SOURCE_DIR}/mscore/importxmlfirstpass.cpp
